all: fix the link error with Go 1.24.3 (#312)
This addresses the issue golang/go#73617.

Closes #313

+++ internal/fakecgo/gen.go
@@ -76,7 +76,7 @@ call5({{.Name}}ABI0,
 {{end}}
 {{- range . }}
 //go:linkname _{{.Name}} _{{.Name}}
-var _{{.Name}} uintptr
+var _{{.Name}} uint8
 var {{.Name}}ABI0 = uintptr(unsafe.Pointer(&_{{.Name}}))
 {{ end }}
 `
